Add BinaryHeap, Memo, and StateMachine, fixup previous caught bugs
