+++ core/async.ts
@@ -143,3 +143,4 @@ export function pinkyPromise<T>(): [
     // the assignment to `p`
     return [p, resolve!, reject!]
 }
+
